Add files via upload
Abstract - The use of photovoltaic (PV) systems is expanding globally as a result of the rising need for energy. Since PV systems are not linear, unique techniques are used to harvest the most power possible. Due to variations in solar intensity, the energy produced by PV panels when sunlight is directly absorbed is not constant. High-performance artificial intelligence (AI) and natural selection techniques can be employed as maximum power point-tracking techniques. Maximum power point tracking (MPPT) is used to get the most electricity possible from the solar panels. In this research, to reduce the energy loss supplied by the PV module and to increase energy efficiency, we will examine and contrast the performances of MPPT approaches control systems composed of a PV solar system under various irradiation situations. This control is implemented using the MPPT under the MATLAB/Simulink environment, which is based on the Perturb and Observe algorithm (P&O), Fuzzy Logic algorithm (FL), and NARMA-L2 Controller algorithm. These three algorithms are analyzed and their functional analysis has been done to find the most optimum technique.